Updated some comments.

+++ inc/BitFunnel/Index/IFactSet.h
@@ -1,39 +1,42 @@
 #pragma once
 
-#include <inttypes.h> // For uint64_t.
+#include <stddef.h>
 
 namespace BitFunnel
 {
     //*************************************************************************
     //
-    // IFactSet is an abstract base class or an interface for classes that 
+    // IFactSet is an abstract base class or an interface for classes that
     // maintain a collection of FactInfo objects that describe document facts.
-    // A document fact represents a predicate or assertion that is either true 
-    // or false for a given document, e.g. document is in top 20% static rank, 
+    // A document fact represents a predicate or assertion that is either true
+    // or false for a given document, e.g. document is in top 20% static rank,
     // or document was indexed before 5pm.
     //
-    // Document facts are typically implemented as private, rank 0 rows in the 
+    // Document facts are typically implemented as private, rank 0 rows in the
     // index. Document facts may be mutable or immutable. An immutable fact for
-    // a particular document is considered constant for the life of the 
+    // a particular document is considered constant for the life of the
     // document. A mutable fact may be changed at any time.
     //
     // The IFactSet is used to configure the TermTable to be able to set and
-    // clear (clear - only for mutable facts) facts about documents in the 
-    // index and be able to alter queries to include the facts in the query 
+    // clear (clear - only for mutable facts) facts about documents in the
+    // index and be able to alter queries to include the facts in the query
     // plan.
     //
     //*************************************************************************
-
-    typedef uint64_t FactHandle;
+    // TODO: REVIEW:
+    //   Should FactHandle be a member of IFactSet?
+    //   Should IIngestionIndex provide a DefineFact() method?
+    //   Should IFactSet be publicly visible?
+    typedef size_t FactHandle;
 
     class IFactSet
     {
     public:
         virtual ~IFactSet() {};
 
-        // Allocates a fact with sequentially numbered FactHandles starting 
+        // Allocates a fact with sequentially numbered FactHandles starting
         // at 1.
-        virtual FactHandle DefineFact(char const * friendlyName, 
+        virtual FactHandle DefineFact(char const * friendlyName,
                                       bool isMutable) = 0;
 
         // Returns the number of facts defined in the IFactSet.
